Localize time habit is completed
Why:
When a user completes a task it is considered complete for the day. We
would like for the time of day to be localized to the users time zone.

This commit:
Gets the local time zone on initialization and uses it to determine
whether or not the habit was completed that day. We also factored the
habit model into its own module and added some basic specs.